update run config for Kr full sims

+++ NEXT100/job/NEXT100_Kr83m_Full.sh
@@ -55,7 +55,13 @@ city hypathia hypathiaTemplate.conf -i NEXT100_Kr83m_Full_buffy_${JOBID}.h5    -
 city dorothea dorotheaTemplate.conf -i NEXT100_Kr83m_Full_hypathia_${JOBID}.h5 -o NEXT100_Kr83m_Full_dorothea_${JOBID}.h5
 
 rm NEXT100_Kr83m_Full.h5
-rm NEXT100_Kr83m_Full_buffy_${JOBID}.h5
+
+# Only keep first 1000 files for validation purposes
+if (( JOBID > 1000 )); then
+    rm NEXT100_Kr83m_Full_buffy_${JOBID}.h5
+    rm NEXT100_Kr83m_Full_hypathia_${JOBID}.h5
+fi
+
 rm *LT*
 rm *PSF*
 rm *map*
